+++ package.json
@@ -7,12 +7,12 @@
     "build": "NODE_ENV=production  ./utils/convertToes5.sh && webpack --colors --display-error-details --config webpack/webpack.prod.js --env.theme=${theme}",
     "build:theme": "NODE_ENV=production webpack --colors --display-error-details --config webpack/webpack.theme.prod.js --env.theme=${theme}",
     "build:admin": "NODE_ENV=production webpack --colors --display-error-details --config webpack/webpack.admin.prod.js",
-    "clean": "rm -rf data && rm -rf public/uploads",
     "seed": "node ./src/api/seed",
     "preinstall": "./utils/preinstall.sh",
     "postinstall": "./utils/postinstall.sh",
     "dev": "NODE_ENV=dev concurrently --kill-others \"node ./src/server.js --profile --json\" \"webpack --config webpack/webpack.dev.js --env.theme=${theme}\"",
     "prod": "concurrently --kill-others \"node ./dist/server.js\"",
+    "start": "yarn prod",
     "prodPm2": "./forever.sh",
     "static": "NODE_ENV=production & MODE=static concurrently \"node ./server.js\"",
     "test": "NODE_ENV=test jest --runInBand",
@@ -45,6 +45,7 @@
     "bcryptjs": "^2.4.3",
     "body-parser": "^1.19.0",
     "compression": "^1.7.4",
+    "concurrently": "^4.1.2",
     "cors": "^2.8.5",
     "disqus-react": "^1.0.6",
     "express": "^4.17.1",
@@ -111,7 +112,6 @@
     "babel-plugin-styled-components": "^1.10.6",
     "babel-preset-react-hmre": "^1.1.1",
     "better-npm-run": "0.1.1",
-    "concurrently": "^4.1.2",
     "copy-dir": "^1.2.0",
     "copy-webpack-plugin": "^5.0.4",
     "core-js": "3",
