+++ src/client/server/html.js
@@ -35,9 +35,9 @@ module.exports.getHtml = (
   const scripts = isStatic ? "" : prepareScriptTags(bundles);
 
   // get the styles only in production. for dev, it will be injected by webpack
-  const styleLinks = !isDev
-    ? prepareStyleTags(theme + "/dist/client.min.css")
-    : "";
+  const styleLinks = isDev
+    ? ""
+    : prepareStyleTags(theme + "/dist/client.min.css");
 
   // check if the theme has defined any html template
   const themeTemplateBuffer = util.getThemeFileContents(theme, "template.tpl");
