+++ utils/preinstall.sh
@@ -5,8 +5,8 @@ printf "$npm_execpath" | grep -q "yarn\.js$" ||
 \n⚠️ Use yarn not npm! ⚠️\n
 ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~ ~" && exit 1)
 
-echo "Installing the default theme - Hugo"
-if [ ! -d src/client/themes/hugo ]; then
-    git clone https://github.com/letterpad/theme-hugo.git client/themes/hugo --depth 1
+if [ ! -f src/client/themes/hugo/package.json ]; then
+    echo "Installing the default theme - Hugo"
+    git clone https://github.com/letterpad/theme-hugo.git src/client/themes/hugo --depth 1
 fi
